requirements.txt
fastapi
uvicorn
matplotlib
numpy
opencv-python
Pillow
python-multipart
pytesseract
torch
torchvision